app/requirements.txt
streamlit>=1.30
plotly>=5.18
pandas>=2.1
numpy>=1.26
statsmodels>=0.14
oracledb>=2.0
anthropic>=0.40
yfinance>=0.2.31
apscheduler>=3.10
python-dotenv>=1.0
